# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pelayanan Publik API",
        "description": "API Backend untuk Aplikasi Pelayanan Publik",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/layanan": {
            "get": {
                "tags": [
                    "Layanan"
                ],
                "summary": "Get all layanan publik",
                "operationId": "9f105eb8582bf1cff15e81aa8fa8216c",
                "parameters": [
                    {
                        "name": "jenis",
                        "in": "query",
                        "description": "Filter by jenis layanan ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "online",
                        "in": "query",
                        "description": "Filter online layanan only",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search layanan",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of layanan"
                    }
                }
            }
        },
        "/layanan/{id}": {
            "get": {
                "tags": [
                    "Layanan"
                ],
                "summary": "Get single layanan",
                "operationId": "f0ab4c34c08a0ed2ad482b007dce25cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Layanan details"
                    }
                }
            }
        },
        "/admin/layanan": {
            "post": {
                "tags": [
                    "Layanan"
                ],
                "summary": "Create new layanan",
                "operationId": "4a7415b41bd29e747d2b6a173f1b3497",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_layanan",
                                    "deskripsi",
                                    "syarat_prosedur",
                                    "id_jenis_layanan"
                                ],
                                "properties": {
                                    "nama_layanan": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "syarat_prosedur": {
                                        "type": "string"
                                    },
                                    "id_jenis_layanan": {
                                        "type": "integer"
                                    },
                                    "durasi_pelayanan": {
                                        "type": "string"
                                    },
                                    "biaya": {
                                        "type": "number"
                                    },
                                    "is_online": {
                                        "type": "boolean"
                                    },
                                    "link_online": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Layanan created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/layanan/{id}": {
            "put": {
                "tags": [
                    "Layanan"
                ],
                "summary": "Update layanan",
                "operationId": "e04572f4f80391295efdadde7f60d26b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Layanan updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Layanan"
                ],
                "summary": "Delete layanan",
                "operationId": "76ef83deb672e4510f3bc1d4ea8df9f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Layanan deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}